Fix Black formatting and improve component robustness
- Fix Black formatting issue in LLM agent temperature assignment
- Improve data type handling in VuManChu MFI calculation with explicit float64 conversion
- Add order manager method for adding existing orders to support testing scenarios
- Fix validator rationale truncation to prevent schema validation errors
- Update integration test assertions to match actual indicator output structure

+++ bot/indicators/vumanchu.py
@@ -1379,10 +1379,10 @@ class CipherB:
                 # Calculate Money Flow Index (MFI) using pandas-ta
                 logger.debug("Calculating MFI")
                 mfi_values = ta.mfi(
-                    df["high"],
-                    df["low"],
-                    df["close"],
-                    df["volume"],
+                    df["high"].astype("float64"),
+                    df["low"].astype("float64"),
+                    df["close"].astype("float64"),
+                    df["volume"].astype("float64"),
                     length=self.mfi_length,
                 )
                 df["money_flow"] = (
